+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add device name for mobile integration tests
 - Added new FAQs and updated Readme to include new demos for PSTN calling and live events
+- Add test report URL for mobile integration tests
 
 ### Changed
 - Fix title for FAQ guide